% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_operator.R
\name{check_operator}
\alias{check_operator}
\title{Check validity of operator and values}
\usage{
check_operator(operator, values)
}
\arguments{
\item{operator}{The operator to be checked.}

\item{values}{The values to be checked.}
}
\value{
NULL (no explicit return value).
}
\description{
This function checks the validity of the operator and values used in a condition.
}
