% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_id_introduction_steps.R
\name{discover_id_introduction_steps}
\alias{discover_id_introduction_steps}
\title{Discover ID Introduction Steps}
\usage{
discover_id_introduction_steps(conn, full_path_dataframe)
}
\arguments{
\item{conn}{The connection object or database connection string.}

\item{full_path_dataframe}{A data frame representing the full join path, including columns: "table_to_join",
"method", and "common_var".}
}
\value{
A data frame with information about newly discovered IDs and the corresponding join step.
}
\description{
This function identifies the steps in a join path where new IDs are introduced, allowing you to
determine at which join steps each ID variable is added to the query. It returns a data frame with
information about newly discovered IDs and the corresponding join step in the path.
}
