% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_valid_sql.R
\name{make_valid_sql}
\alias{make_valid_sql}
\title{Create a valid SQL statement based on variable, operator, and values}
\usage{
make_valid_sql(conn, variable, operator, values)
}
\arguments{
\item{conn}{The connection object or database connection string.}

\item{variable}{The variable for which the SQL statement is created.}

\item{operator}{The operator to be used in the SQL statement.}

\item{values}{The values to be used in the SQL statement.}
}
\value{
A character string representing the valid SQL statement.
}
\description{
This function creates a valid SQL statement based on the specified variable, operator, and values.
It handles different operators such as "greater", "less", "equal", and "between".
}
