% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompute_table_join_paths.R
\name{precompute_table_join_paths}
\alias{precompute_table_join_paths}
\title{Precompute Table Join Paths}
\usage{
precompute_table_join_paths(conn, input_table = NULL, relevant_tables = NULL)
}
\arguments{
\item{conn}{The connection object or database connection string.}

\item{input_table}{The table from which the join path is computed.}

\item{relevant_tables}{A vector of tables that are relevant to the query.}
}
\value{
A list of join paths for each table in the database.
}
\description{
This function precomputes join paths for all tables in a given database using a combination of forward and backward joins.
It generates a list of data frames representing the join paths for each table, including information about tables to join,
walk approaches (forward or backward), and common variables used for joining.
}
