% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_id_name_from_table.R
\name{return_id_name_from_table}
\alias{return_id_name_from_table}
\title{Return ID column name from table name}
\usage{
return_id_name_from_table(table_name)
}
\arguments{
\item{table_name}{The name of the table.}
}
\value{
The generated ID column name.
}
\description{
This function generates the ID column name based on the provided table name.
It replaces the "table" suffix with "id" to obtain the ID column name.
}
