% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_table_name_from_id.R
\name{return_table_name_from_id}
\alias{return_table_name_from_id}
\title{Return table name from ID column name}
\usage{
return_table_name_from_id(id_name)
}
\arguments{
\item{id_name}{The name of the ID column.}
}
\value{
The generated table name.
}
\description{
This function generates the table name based on the provided ID column name.
It replaces the "id" suffix with "table" to obtain the table name.
}
